requirements.txt
# Core ML and Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
keras>=2.13.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0

# GenAI Integration
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Development
jupyter>=1.0.0
notebook>=6.5.0

# Additional dependencies for Python 3.13 compatibility
setuptools>=65.0.0
wheel>=0.40.0
